+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli.Contract/MCPToolInterface.cs
@@ -1,0 +1,11 @@
+using Microsoft.AspNetCore.Routing;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace Azure.Sdk.Tools.Cli.Contract
+{
+    // todo: populate this instead of overrelying on abstract class
+    public interface MCPToolInterface
+    {
+    }
+}