+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli/Commands/SharedOptions.cs
@@ -1,0 +1,80 @@
+using System.CommandLine;
+using System.CommandLine.Parsing;
+using Azure.Sdk.Tools.Cli.Contract;
+using System.IO.Enumeration;
+using System.Reflection;
+
+namespace Azure.Sdk.Tools.Cli.Commands
+{
+    public static class SharedOptions
+    {
+        public static Option<string> ToolOption = new Option<string>("--tools")
+        {
+            Description = "If provided, the tools server will only respond to CLI or MCP server requests for tools named the same as provided in this option. Glob matching is honored.",
+            IsRequired = false,
+        };
+
+        public static List<string> GetToolsFromArgs(string[] args)
+        {
+            var root = new RootCommand
+            {
+                TreatUnmatchedTokensAsErrors = false
+            };
+            root.AddOption(ToolOption);
+
+            var parser = new Parser(root);
+            var result = parser.Parse(args);
+
+            var raw = result.GetValueForOption(ToolOption);
+            if (string.IsNullOrWhiteSpace(raw))
+                return new List<string>();
+
+            return raw
+                .Split(',', StringSplitOptions.RemoveEmptyEntries | StringSplitOptions.TrimEntries)
+                .Select(s => s.ToLowerInvariant())
+                .ToList();
+        }
+
+
+        public static List<Type> GetFilteredToolTypes(string[] args)
+        {
+            var toolMatchList = SharedOptions.GetToolsFromArgs(args);
+            List<Type> toolsList;
+
+            if (toolMatchList.Count > 0)
+            {
+                toolsList = AppDomain.CurrentDomain
+                             .GetAssemblies()
+                             .SelectMany(a => SafeGetTypes(a))
+                             .Where(t => !t.IsAbstract &&
+                             typeof(MCPTool).IsAssignableFrom(t))
+                             .Where(t => toolMatchList.Any(x => FileSystemName.MatchesSimpleExpression(x, t.Name) || t.Name.StartsWith("HostServer")))
+                             .ToList();
+            }
+            else
+            {
+                // defaults to everything
+                toolsList = AppDomain.CurrentDomain
+                             .GetAssemblies()
+                             .SelectMany(a => SafeGetTypes(a))
+                             .Where(t => !t.IsAbstract &&
+                             typeof(MCPTool).IsAssignableFrom(t))
+                             .ToList();
+            }
+
+            return toolsList;
+        }
+
+        public static IEnumerable<Type> SafeGetTypes(Assembly asm)
+        {
+            try
+            {
+                return asm.GetTypes();
+            }
+            catch (ReflectionTypeLoadException ex)
+            {
+                return ex.Types!.Where(t => t != null)!;
+            }
+        }
+    }
+}